server/requirements.txt
flask
aiortc
aiohttp
aioflask
flask-socketio
flask-cors
eventlet
mediapipe
opencv-python==4.5.5.64
